+++ src/main/java/com/programmers/bucketback/domains/bucket/application/BucketRemover.java
@@ -2,7 +2,10 @@ package com.programmers.bucketback.domains.bucket.application;
 
 import org.springframework.stereotype.Component;
 
+import com.programmers.bucketback.domains.bucket.domain.Bucket;
 import com.programmers.bucketback.domains.bucket.repository.BucketRepository;
+import com.programmers.bucketback.global.error.exception.EntityNotFoundException;
+import com.programmers.bucketback.global.error.exception.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -14,6 +17,9 @@ public class BucketRemover {
 
 	/** 버킷 삭제 */
 	public void remove(Long bucketId) {
-		bucketRepository.deleteById(bucketId);
+		bucketRepository.delete(
+			bucketRepository.findById(bucketId)
+				.orElseThrow(() -> new EntityNotFoundException(ErrorCode.BUCKET_NOT_FOUND))
+		);
 	}
 }
